requirements.txt
# acados_template.egg==info
casadi==3.6.6
jax==0.4.30
jaxopt==0.8.3
matplotlib==3.9.2
numpy==2.0.0
ruamel.yaml==0.18.5
scikit_learn==1.5.1
scipy==1.13.1
Shapely==2.0.6
torch==2.4.0
tueplots==0.0.15
